readOnly tests for multiple files.
readOnly test now reads all the files passed with -i parameter.
It will read them individually in datasets and then make a
union dataset out of them. -a actions like count, save,
or collect are executed on the union dataset

WARNING: schema of off the files should be the same!