Introducing input/output formats and options
inputFormat (-if) and options (-ifo) are what is passed to
spark readers. The default is parquet without options.

outputFormat (-of) and options (-ofo) are what is passed to
spark writers. The default is parquet, and compression
disabled (-ofo has <"compression","none">).

These parameters are respected in all the tests now.